# creditsoft/openapi.yaml
openapi: 3.1.0
info:
  title: CreditSoft Local Partner API
  version: 1.0.0
  description: |
    Local-first API for trusted lead-management systems and the CreditSoft website portal.
    Authenticate every write and read request with a bearer token generated inside the intranet settings.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CreditSoftToken
  schemas:
...
    ClientCreate:
      type: object
      required:
        - first_name
        - last_name
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address_line_1:
          type: string
        address_line_2:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        current_score:
          type: integer
          minimum: 300
          maximum: 850
        status:
          type: string
          example: lead
        goals:
          type: string
        external_reference:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    ClientDocumentCreate:
      type: object
      required:
        - title
      properties:
        reporting_cycle_id:
          type:
            - integer
            - "null"
        title:
          type: string
        category:
          type:
            - string
            - "null"
        notes:
          type:
            - string
            - "null"
        portal_visible:
          type:
            - boolean
            - "null"
        document_quality:
          type:
            - object
            - "null"
          description: Optional client-side capture review generated by the portal before upload.
          additionalProperties: true
          properties:
            status:
              type: string
              enum:
                - accepted
                - retake_recommended
            score:
              type: integer
              minimum: 0
              maximum: 100
            warnings:
              type: array
              items:
                type: object
                additionalProperties: true
                properties:
                  code:
                    type: string
                  message:
                    type: string
            dimensions:
              type: object
              additionalProperties: false
              properties:
                width:
                  type: integer
                  minimum: 0
                height:
                  type: integer
                  minimum: 0
            orientation:
              type: string
            document_frame_ratio:
              type:
                - number
                - "null"
        document_file:
          type: string
          format: binary
...
    BrowserCompanionCapture:
      type: object
      properties:
        client_cuid:
          type:
            - string
            - "null"
        client_lookup:
          type:
            - object
            - "null"
          additionalProperties: true
        client_email:
          type:
            - string
            - "null"
          format: email
        client_name:
          type:
            - string
            - "null"
        client_first_name:
          type:
            - string
            - "null"
        client_last_name:
          type:
            - string
            - "null"
        create_client_if_missing:
          type:
            - boolean
            - "null"
        provider_key:
          type:
            - string
            - "null"
        reporting_cycle_id:
          type:
            - integer
            - "null"
        reporting_cycle_label:
          type:
            - string
            - "null"
        cycle_label:
          type:
            - string
            - "null"
        source_type:
          type:
            - string
            - "null"
          enum:
            - dom_capture
            - browser_capture
            - companion_capture
            - safari_webarchive
        browser_name:
          type:
            - string
            - "null"
        page_title:
          type:
            - string
            - "null"
        page_url:
          type:
            - string
            - "null"
        html:
          type:
            - string
            - "null"
        dom_html:
          type:
            - string
            - "null"
        operator_note:
          type:
            - string
            - "null"
        selection_text:
          type:
            - string
            - "null"
        package_name:
          type:
            - string
            - "null"
        user_agent:
          type:
            - string
            - "null"
        captured_at:
          type:
            - string
            - "null"
          format: date-time
        capture_file:
          type: string
          format: binary
paths:
...
  /clients:
    post:
      summary: Create a local lead/client record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        "201":
          description: Client created
...
  /clients/{clientCuid}/documents:
    get:
      summary: List portal-safe documents uploaded for a client
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client documents list
    post:
      summary: Upload a client document with audit logging
      parameters:
        - in: path
          name: clientCuid
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ClientDocumentCreate'
      responses:
        "201":
          description: Client document created
...
  /browser-companion/intake:
    post:
      summary: Resolve a client by lookup or provider-ready queue and ingest a browser companion capture into the right cycle
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BrowserCompanionCapture'
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserCompanionCapture'
      responses:
        "201":
          description: Browser companion capture created